+++ src/main/java/org/example/Phone.java
@@ -4,7 +4,7 @@ public class Phone extends TouchScreen {
     private int mainCamera;
     private int frontCamera;
 
-    public Phone(String trademark, String model, int price, int ram, float diagonal, int battery, float weight, int mainCamera, int frontCamera) {
+    public Phone(String trademark, String model, float price, int ram, float diagonal, int battery, float weight, int mainCamera, int frontCamera) {
         super(trademark, model, price, ram, diagonal, battery, weight);
         this.mainCamera = mainCamera;
         this.frontCamera = frontCamera;
