+++ src/main/java/org/example/Tablet.java
@@ -5,7 +5,7 @@ import jdk.jfr.Description;
 public class Tablet extends TouchScreen{
     private final float cashbackCoef = 0.003f;
 
-    public Tablet(String trademark, String model, int price, int ram, float diagonal, int battery, float weight) {
+    public Tablet(String trademark, String model, float price, int ram, float diagonal, int battery, float weight) {
         super(trademark, model, price, ram, diagonal, battery, weight);
     }
